+++ MyList.Domain/Common/Models/ContentModels/Game.cs
@@ -11,8 +11,8 @@ namespace MyList.Domain.Common.Models.ContentModels
         public string? Description { get; set; }
         public string? PictureURL { get; set; }
         public ICollection<GameTag>? Tags { get; set; } = new List<GameTag>();
-        public int? UserScore { get; set; } //enum
-        public int? GlobalScore { get; set; } //enum
+        public float? UserScore { get; set; } //enum
+        public float? GlobalScore { get; set; } //enum
         public string? UserStatus { get; set; } //enum
         public string? GlobalStatus { get; set; } //enum
         public string? GameStudio { get; set; }
